refactor fixation and timerstore away from common
