+++ x/dualstaking/types/expected_keepers.go
@@ -6,9 +6,10 @@ import (
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/lavanet/lava/common"
 	epochstoragetypes "github.com/lavanet/lava/x/epochstorage/types"
+	"github.com/lavanet/lava/x/fixationstore"
 	spectypes "github.com/lavanet/lava/x/spec/types"
+	"github.com/lavanet/lava/x/timerstore"
 )
 
 // AccountKeeper defines the expected account keeper used for simulations (noalias)
@@ -48,9 +49,9 @@ type StakingKeeper interface {
 }
 
 type FixationStoreKeeper interface {
-	NewFixationStore(storeKey storetypes.StoreKey, prefix string) *common.FixationStore
+	NewFixationStore(storeKey storetypes.StoreKey, prefix string) *fixationstore.FixationStore
 }
 
 type TimerStoreKeeper interface {
-	NewTimerStore(storeKey storetypes.StoreKey, prefix string) *common.TimerStore
+	NewTimerStore(storeKey storetypes.StoreKey, prefix string) *timerstore.TimerStore
 }
