+++ x/fixationstore/fixationstore_migrations.go
@@ -1,4 +1,4 @@
-package common
+package fixationstore
 
 import (
 	"fmt"
