+++ x/fixationstore/keeper.go
@@ -4,7 +4,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/lavanet/lava/common"
 	"github.com/lavanet/lava/x/timerstore"
 )
 
@@ -18,14 +17,14 @@ func NewKeeper(cdc codec.BinaryCodec, tsKeeper *timerstore.Keeper) *Keeper {
 // Keeper is the fixationstore keeper. The keeper retains all the fixation stores used by modules,
 // it also manages their lifecycle.
 type Keeper struct {
-	fixationsStores []*common.FixationStore
+	fixationsStores []*FixationStore
 	ts              *timerstore.Keeper
 	cdc             codec.BinaryCodec
 }
 
-func (k *Keeper) NewFixationStore(storeKey storetypes.StoreKey, prefix string) *common.FixationStore {
+func (k *Keeper) NewFixationStore(storeKey storetypes.StoreKey, prefix string) *FixationStore {
 	ts := k.ts.NewTimerStore(storeKey, prefix)
-	fs := common.NewFixationStore(storeKey, k.cdc, prefix, ts)
+	fs := NewFixationStore(storeKey, k.cdc, prefix, ts)
 	k.fixationsStores = append(k.fixationsStores, fs)
 	return fs
 }
