+++ x/pairing/keeper/keeper.go
@@ -4,13 +4,14 @@ import (
 	"fmt"
 
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
+	"github.com/lavanet/lava/x/fixationstore"
+	"github.com/lavanet/lava/x/timerstore"
 
 	"github.com/cometbft/cometbft/libs/log"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-	"github.com/lavanet/lava/common"
 	commontypes "github.com/lavanet/lava/common/types"
 	"github.com/lavanet/lava/x/pairing/types"
 )
@@ -29,8 +30,8 @@ type (
 		projectsKeeper     types.ProjectsKeeper
 		subscriptionKeeper types.SubscriptionKeeper
 		planKeeper         types.PlanKeeper
-		badgeTimerStore    common.TimerStore
-		providerQosFS      common.FixationStore
+		badgeTimerStore    timerstore.TimerStore
+		providerQosFS      fixationstore.FixationStore
 		downtimeKeeper     types.DowntimeKeeper
 		dualStakingKeeper  types.DualStakingKeeper
 	}
