+++ x/pairing/types/expected_keepers.go
@@ -5,12 +5,13 @@ import (
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/lavanet/lava/common"
 	epochstoragetypes "github.com/lavanet/lava/x/epochstorage/types"
+	"github.com/lavanet/lava/x/fixationstore"
 	planstypes "github.com/lavanet/lava/x/plans/types"
 	projectstypes "github.com/lavanet/lava/x/projects/types"
 	spectypes "github.com/lavanet/lava/x/spec/types"
 	subscriptiontypes "github.com/lavanet/lava/x/subscription/types"
+	"github.com/lavanet/lava/x/timerstore"
 )
 
 type SpecKeeper interface {
@@ -95,9 +96,9 @@ type DualStakingKeeper interface {
 }
 
 type FixationStoreKeeper interface {
-	NewFixationStore(storeKey storetypes.StoreKey, prefix string) *common.FixationStore
+	NewFixationStore(storeKey storetypes.StoreKey, prefix string) *fixationstore.FixationStore
 }
 
 type TimerStoreKeeper interface {
-	NewTimerStore(storeKey storetypes.StoreKey, prefix string) *common.TimerStore
+	NewTimerStore(storeKey storetypes.StoreKey, prefix string) *timerstore.TimerStore
 }
