+++ x/plans/keeper/keeper.go
@@ -4,13 +4,13 @@ import (
 	"fmt"
 
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
+	"github.com/lavanet/lava/x/fixationstore"
 
 	"github.com/cometbft/cometbft/libs/log"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
-	"github.com/lavanet/lava/common"
 	commontypes "github.com/lavanet/lava/common/types"
 	"github.com/lavanet/lava/x/plans/types"
 )
@@ -23,7 +23,7 @@ type (
 		epochstorageKeeper types.EpochStorageKeeper
 		specKeeper         types.SpecKeeper
 
-		plansFS common.FixationStore
+		plansFS fixationstore.FixationStore
 	}
 )
