+++ x/projects/types/expected_keepers.go
@@ -3,7 +3,7 @@ package types
 import (
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/lavanet/lava/common"
+	"github.com/lavanet/lava/x/fixationstore"
 )
 
 type EpochStorageKeeper interface {
@@ -12,5 +12,5 @@ type EpochStorageKeeper interface {
 }
 
 type FixationStoreKeeper interface {
-	NewFixationStore(storeKey storetypes.StoreKey, prefix string) *common.FixationStore
+	NewFixationStore(storeKey storetypes.StoreKey, prefix string) *fixationstore.FixationStore
 }
