+++ x/subscription/types/genesis.go
@@ -1,6 +1,8 @@
 package types
 
-import common "github.com/lavanet/lava/common"
+import (
+	"github.com/lavanet/lava/x/fixationstore"
+)
 
 // this line is used by starport scaffolding # genesis/types/import
 
@@ -12,7 +14,7 @@ func DefaultGenesis() *GenesisState {
 	return &GenesisState{
 		// this line is used by starport scaffolding # genesis/types/default
 		Params: DefaultParams(),
-		SubsFS: *common.DefaultGenesis(),
+		SubsFS: *fixationstore.DefaultGenesis(),
 	}
 }
