+++ x/timerstore/keeper.go
@@ -4,7 +4,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/lavanet/lava/common"
 )
 
 func NewKeeper(cdc codec.BinaryCodec) *Keeper {
@@ -16,12 +15,12 @@ func NewKeeper(cdc codec.BinaryCodec) *Keeper {
 // Keeper is the timerstore keeper. The keeper retains all the fixation stores used by modules,
 // it also manages their lifecycle.
 type Keeper struct {
-	timerStores []*common.TimerStore
+	timerStores []*TimerStore
 	cdc         codec.BinaryCodec
 }
 
-func (k *Keeper) NewTimerStore(storeKey storetypes.StoreKey, prefix string) *common.TimerStore {
-	ts := common.NewTimerStore(storeKey, k.cdc, prefix)
+func (k *Keeper) NewTimerStore(storeKey storetypes.StoreKey, prefix string) *TimerStore {
+	ts := NewTimerStore(storeKey, k.cdc, prefix)
 	k.timerStores = append(k.timerStores, ts)
 	return ts
 }
